# openapi.yaml
{
  "openapi":"3.1.0",
  "info":{
    "title":"OpenAPI definition",
    "version":"v0"
  },
  "servers":[
    {
      "url":"http://localhost:8080",
      "description":"Generated server url"
    }
  ],
  "tags":[
    {
      "name":"Clicker",
      "description":"Endpoints for clicker"
    },
    {
      "name":"User",
      "description":"Endpoints for user"
    }
  ],
  "paths":{
    "/api/register":{
      "post":{
        "tags":[
          "User"
        ],
        "summary":"Register endpoint",
        "description":"Here is the registration endpoint, it should be returned if you were able to register or not!",
        "operationId":"userRegister",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/UserRequestDTO"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"Ok",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/UserRegisterResponseDTO"
                }
              }
            }
          },
          "404":{
            "description":"User suplied not found",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/login":{
      "post":{
        "tags":[
          "User"
        ],
        "summary":"Login endpoint",
        "description":"Here is the login endpoint, it should be return if you were able to log or not!",
        "operationId":"userLogin",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/UserRequestDTO"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"Ok",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/UserLoginResponseDTO"
                }
              }
            }
          },
          "404":{
            "description":"User suplied not found",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/clicker":{
      "post":{
        "tags":[
          "Clicker"
        ],
        "summary":"Principal endpoint for clicker",
        "description":"Here is the principal endpoint where should be return if you won or you lose.",
        "operationId":"clicked",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "$ref":"#/components/schemas/ClickerRequestDTO"
              }
            }
          },
          "required":true
        },
        "responses":{
          "200":{
            "description":"Ok",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/ClickerResponseDTO"
                }
              }
            }
          },
          "404":{
            "description":"User suplied not found",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/ErrorResponseDTO"
                }
              }
            }
          },
          "403":{
            "description":"Unathorized",
            "content":{
              "*/*":{
                "schema":{
                  "$ref":"#/components/schemas/ClickerResponseDTO"
                }
              }
            }
          }
        },
        "security":[
          {
            "bearerAuth":[

            ]
          }
        ]
      }
    },
    "/api/leadboard":{
      "get":{
        "tags":[
          "User"
        ],
        "summary":"Leadboard endpoint",
        "description":"Here is the leadboard endpoint, it should be a list with all winner and your attempts to won the game.",
        "operationId":"getLeadboard",
        "responses":{
          "200":{
            "description":"Ok",
            "content":{
              "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/LeadboardResponse"
                }
              }
            }
          }
        },
        "security":[
          {
            "bearerAuth":[

            ]
          }
        ]
      }
    }
  },
  "components":{
    "schemas":{
      "UserRequestDTO":{
        "type":"object",
        "properties":{
          "username":{
            "type":"string"
          },
          "email":{
            "type":"string"
          },
          "password":{
            "type":"string"
          }
        }
      },
      "UserRegisterResponseDTO":{
        "type":"object",
        "properties":{
          "name":{
            "type":"string"
          },
          "email":{
            "type":"string"
          }
        }
      },
      "ErrorResponseDTO":{
        "type":"object",
        "properties":{
          "status":{
            "type":"integer",
            "format":"int32"
          },
          "message":{
            "type":"string"
          },
          "field":{
            "type":"string"
          },
          "time":{
            "type":"string",
            "format":"date-time"
          }
        }
      },
      "UserLoginResponseDTO":{
        "type":"object",
        "properties":{
          "username":{
            "type":"string"
          },
          "email":{
            "type":"string"
          },
          "randomNumber":{
            "type":"integer",
            "format":"int32"
          },
          "avaliableClick":{
            "type":"integer",
            "format":"int32"
          },
          "token":{
            "type":"string"
          }
        }
      },
      "ClickerRequestDTO":{
        "type":"object",
        "properties":{
          "username":{
            "type":"string"
          }
        }
      },
      "ClickerResponseDTO":{
        "type":"object",
        "properties":{
          "message":{
            "type":"string"
          }
        }
      },
      "LeadboardResponse":{
        "type":"object",
        "properties":{
          "username":{
            "type":"string"
          },
          "randomNumber":{
            "type":"integer",
            "format":"int32"
          },
          "attempts":{
            "type":"integer",
            "format":"int32"
          },
          "winner":{
            "type":"boolean"
          }
        }
      }
    },
    "securitySchemes":{
      "bearerAuth":{
        "type":"http",
        "scheme":"bearer",
        "bearerFormat":"JWT"
      }
    }
  }
}